mixnet::core

Struct Mixnet

pub struct Mixnet<X> { /* private fields */ }

Mixnet core state. X is the type of the extra data stored for each mixnode (Mixnode::extra).

Implementations

impl<X> Mixnet<X>

pub fn new(config: Config) -> Self

Create a new Mixnet.

pub fn session_status(&self) -> SessionStatus

Returns the current session index and phase.

pub fn set_session_status(&mut self, session_status: SessionStatus)

Sets the current session index and phase. The current and previous mixnodes may need to be provided after calling this; see maybe_set_mixnodes.

pub fn maybe_set_mixnodes( &mut self, rel_session_index: RelSessionIndex, mixnodes: &mut dyn FnMut() -> Result<Vec<Mixnode<X>>, MixnodesErr>, )

Sets the mixnodes for the specified session, if they are needed. If mixnodes() returns Err(MixnodesErr::Permanent), the session slot will be disabled, and later calls to maybe_set_mixnodes for the session will return immediately. If mixnodes() returns Err(MixnodesErr::Transient), the session slot will merely remain empty, and later calls to maybe_set_mixnodes may succeed.

The mixnode peer IDs are used for two things:

Checking for connectivity (they are passed to NetworkStatus::is_connected).

Sending packets (they are put in AddressedPacket::peer_id).

pub fn next_kx_public(&mut self) -> &KxPublic

Returns the key-exchange public key for the next session.

pub fn reserved_peers(&self) -> impl Iterator<Item = &Mixnode<X>>

Returns the mixnodes we should try to maintain connections to.

pub fn handle_packet(&mut self, packet: &Packet) -> Option<Message>

Handle an incoming packet. If the packet completes a message, the message is returned. Otherwise, None is returned.

pub fn next_forward_packet_deadline(&self) -> Option<Instant>

Returns the next instant at which pop_next_forward_packet should be called. None means never.

pub fn pop_next_forward_packet(&mut self) -> Option<AddressedPacket>

Pop and return the packet at the head of the forward packet queue. Returns None if the queue is empty.

pub fn next_authored_packet_delay(&self) -> Option<Duration>

Returns the delay after which pop_next_authored_packet should be called. None means an infinite delay.

pub fn pop_next_authored_packet( &mut self, ns: &dyn NetworkStatus, ) -> Option<AddressedPacket>

Either generate and return a cover packet or pop and return the packet at the head of one of the authored packet queues. May return None if cover packets are disabled, we fail to generate a cover packet, or there are no active sessions (though in the no active sessions case next_authored_packet_delay should return None and so this function should not really be called).

pub fn post_request( &mut self, session_index: SessionIndex, destination_index: &mut Option<MixnodeIndex>, message_id: &MessageId, data: Scattered<'_, u8>, num_surbs: usize, ns: &dyn NetworkStatus, ) -> Result<RequestMetrics, PostErr>

Post a request message. If destination_index is None, a destination mixnode is chosen at random and (on success) its index is written back to destination_index. The message is split into fragments and each fragment is sent over a different path to the destination.

pub fn post_reply( &mut self, surbs: &mut Vec<Surb>, session_index: SessionIndex, message_id: &MessageId, data: Scattered<'_, u8>, ) -> Result<(), PostErr>

Post a reply message using SURBs. The session index must match the session the SURBs were generated for. SURBs are removed from surbs on use.

pub fn take_events(&mut self) -> Events

Clear the event flags. Returns the flags that were cleared.