mixnet

Module request_manager

For more reliable delivery, a RequestManager can be used instead of calling Mixnet::post_request directly. A RequestManager serves as an additional buffer for requests, and will retry posting if requests are not removed within the expected time.

Structs

Config

Request manager configuration.

RequestManager

Request manager state. The user is responsible for calling update_session_status, process_post_queues, and pop_next_retry at the appropriate times to make progress.

Traits

Request

Requests managed by a RequestManager must implement this trait.